operator/requirements.txt
aiohttp==3.6.2
aiojobs==0.2.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
google-auth==1.14.3
idna==2.9
iso8601==0.1.12
isort==4.3.21
kopf==0.26
kubernetes==11.0.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pykube-ng==20.5.0
pylint==2.5.2
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
websocket-client==0.57.0
wrapt==1.12.1
yarl==1.4.2
